requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
matplotlib
mongoengine==0.27.0
nbformat
numpy==1.26.4
pandas==2.2.0
plotly_express
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
scikit-learn==1.4.0
seaborn
scipy==1.12.0
six==1.16.0
spotipy==2.23.0
threadpoolctl==3.3.0
tqdm
tzdata==2024.1
urllib3==2.2.0
yellowbrick
Werkzeug==3.0.1



